cloud/generate_tile_info/requirements.txt
functions_framework==3.5.0
google-cloud-bigquery==3.18.0
google-cloud-storage==2.16.0
pyproj==3.6.1
python-dotenv==1.0.1
Shapely==2.0.3
